+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/utils/Utils.java
@@ -36,7 +36,6 @@ import java.nio.file.Paths;
 import java.util.AbstractMap;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
@@ -234,7 +233,7 @@ public class Utils {
    * Utility method to generate UUIDs.
    * This is taken from Spring Framework's <a href=
    * "https://github.com/spring-projects/spring-framework/blob/a4db0e7448287028d228d46fe7b4df202150958a/spring-core/src/main/java/org/springframework/util/SimpleIdGenerator.java#L35">SimpleIdGenerator</a>
-   * 
+   *
    * @return generated UUID
    */
   public static UUID generateId() {
@@ -292,20 +291,8 @@ public class Utils {
     return str == null || str.isEmpty();
   }
 
-  public static boolean isNotNullOrEmpty(Map<?, ?> map) {
-    return !isNullOrEmpty(map);
-  }
-
-  public static boolean isNullOrEmpty(Map<?, ?> map) {
-    return map == null || map.isEmpty();
-  }
-
-  public static boolean isNotNullOrEmpty(Collection<?> collection) {
-    return !isNullOrEmpty(collection);
-  }
-
-  public static boolean isNullOrEmpty(Collection<?> collection) {
-    return collection == null || collection.isEmpty();
+  public static boolean isNotNullOrEmpty(Map map) {
+    return !(map == null || map.isEmpty());
   }
 
   public static boolean isNotNullOrEmpty(String str) {
@@ -313,11 +300,7 @@ public class Utils {
   }
 
   public static boolean isNotNullOrEmpty(String[] array) {
-    return !isNullOrEmpty(array);
-  }
-
-  public static boolean isNullOrEmpty(String[] array) {
-    return array == null || array.length == 0;
+    return !(array == null || array.length == 0);
   }
 
   public static <T> boolean isNotNull(T... refList) {